MXS-2808: Restart MaxScale on bad cached config
It's possible for the cached configuration to be partially applied when
the error is returned from process_cached_config(). This means that all
objects must be destroyed before an attempt to use the static
configuration is made. A simple way to achieve this is to configure a
special return code that SystemD interprets as a request to restart the
process. This will cause the same cleanup path to be taken that is used in
a normal shutdown. The same exit code can eventually be reused to trigger
a restart if we fail the rollback after a failed config sync with the
cluster. Exiting the process is a cleaner solution than raising a SIGABRT
as it allows the connections to terminate gracefully.

The exact error code is controlled by the RestartForceExitStatus option in
maxscale.service. The value was chosen to be the same that BSD based
systems use for temporary service failures (EX_TEMPFAIL). If MaxScale is
used without SystemD, this will simply appear as a failure to start the
MaxScale process.

Another option would be to iterate the configuration in reverse and
destroy the objects if they exist. This would probably work in most
scenarios and can be added on top of the existing restarting mechanism.

Also extended the error messages that are logged when the application of a
configuration at runtime fails. Until we know what the desired behavior
is, the simplest and least disruptive thing to do is to not crash MaxScale
in the hopes that the problem will be solved by the next configuration
update.