MXS-2808: Crash MaxScale on unknown config state
If the transition from a cached configuration to the cluster configuration
fails in an irreversible way, an attempt to remove the cached
configuration is made. If the file wasn't removed or doesn't exist in the
first place, the process must not restart as it would result in an endless
restart loop.