MXS-2808: Keep original JSON intact
If the JSON is permanently modified during the processing, the checksums
for the configurations will be different depending on whether the node
created the configuration or synced the configuration. As we want to use
the checksum as a quick check of the node's state, it should be the same
for all nodes that successfully apply the configuration.