MXS-3569 Add `line-chart` and show it `chart-container`
A preliminary creation of `line-chart` that reacts on responsive layout.
`chart-container` will contain all supported charts. For now,
only line-chart is added.

Chart configuration will be handled in the following commits
in `visualize-sidebar` to generate `labels` and `datasets`.
The result object will be emitted to `query-page`.
It's then passed down to `chart-container`