Add non-field getters to mxb::Json
This simplifies the use of mxb::Json with JSON Pointers by letting the
object itself be inspected instead of having to first get the parent
object that contains it. This also allows the extraction of array values
that aren't objects.

Removed the is_null() method as nothing was using it. The type() method
provides the same functionality, albeit a more verbose version.