MXS-3595: Fix old listener endpoints
The GET endpoint for the listener of a service did not return the correct
response when the listener was for a different service. This would happen
if one was trying to destroy a listener of a different service than where
it belonged with MaxCtrl. In addition, the error reported by MaxCtrl was
not correctly formatted.

Added a test case that reproduces the problem and verifies that it is
fixed.