MXS-2590: Fix the documentation for the reload command
The command used the older default command method which can be replaced
with demandCommand. The lack of `false` in the description caused the it
to be a normal command, leading to it being printed, instead of a hidden
command.

Also fixed the line wrapping to be the same as in other files.