MXS-3569 Refactor chart tooltips
This adds `objectTooltip` to show chart tooltip
for a single dataset with provided `dataPointObj`

This also measures user's cursor position on the chart
to detect whether tooltip should be aligned on the
left of right.

Also, in `line-chart`, the component is refactored to serve
as a very simple line chart with basic configuration related
to chart style. Other configuration should be done on the parent
component. Also adds `hasVertCrossHair` to `line-chart`, if
this is true, it shows a vertical cross hair while hovering the chart.