MXS-3596: Extend config_sync test
Tests the behavior when the whole cluster is down as well as modifications
done to the database.

Added helper methods that extract the differing parts of JSON. This makes
it easier to see where two JSON values differ without having to manually
process the test output.

The new additions revealed some behavior which might need to be changed in
the future: if a database is restored to an earlier version and all
MaxScale instances are ahead of the database, all attempts at updating the
configuration will fail due to a version mismatch. This requires a manual
step that either deletes the row or updates the version to a new value.