MXS-2808: Fix restart on bad cached config
The way RestartForceExitStatus works with services with Type=forking is
somewhat confusing: the restart is only done if final process, not the
initial process that systemd start, exits with the code found in
RestartForceExitStatus. This means the parent process must exit with the
code 0 and the child must exit with 75.