Optimize MaxCtrl requires
The inquirer library was the biggest problem: it pulled in lots of
dependencies and was by far the slowest module to load. Loading it only
when required cuts down the execution time by roughly a third.

The colors library also slows things down a bit so not loading it in for
the --help output speeds things up by a little bit.