Fix non-TTY command input
The inquirer library doesn't seem to be intended to be used with non-TTY
inputs as it seems to error out when stdin is closed. Even if it did work,
any input that is piped into maxctrl from a non-TTY input should not
output the prompt.

With this change, multiple actions that are loaded from a file can be
performed with minimal overhead. For example, piping a list of 25 `list
servers` commands into maxctrl takes about 0.4 seconds whereas with
individual commands it takes about 5 seconds.