Allow empty value to be set via MaxCtrl
In the INI file a string parameter can be assigned an empty value with
`param=""` as the quotes are eliminated during configuration parsing but
it was impossible do to via MaxCtrl as it refused to accept null or empty
values. Empty values must be allowed to be able to both set and clear them
via MaxCtrl.