MXS-2808: Don't sync config_sync_cluster
Syncing config_sync_cluste causes more problems than it solves which is
why it's not included in the generated configuration. As for the other
cluster parameters, they don't cause problems but it might be wise to not
sync them either.

Whenever config_sync_cluster is disabled at runtime, the configuration it
generates will be identical to the current one which causes the sync to be
skipped. If the first change done on a cluster disabled
config_sync_cluster, it caused the configuration to be sent to the cluster
as m_current_config was empty. To detect this case, the configuration
should be created in ConfigManager::start() if it doesn't exist.