MXS-3596: Test local failures
Covers the case when the write to the disk fails.

The current behavior is to treat the command as failed if the creation of
the cached configuration fails. This could be changed so that if the
cluster operation works but the local caching fails, it would only
generate a warning of some sorts.