MXS-3569 Add minWidth if line-chart is not a linear chart
When x axis is a numeric field, it's linear chart, then it makes
sense to `autoSkip` ticks and the width of the chart is rendered
reponsively. However, when x axis value is string. e.g. date string
all ticks should be visible and the width of the line chart
is determined by the number of datapoints multiplying by 15 px,
but the final value shouldn't be larger than 15000px.