MXS-2808: Failing to save the config is not an error
If the configuration change has been successfully applied and saved in the
cluster, failing to save the cached configuration is no longer treated as
an error. This reduces the effects of local disk problems by relying on
the configuration found in the cluster and avoid the unnecessary rollback
that would otherwise occur.