MXS-2808: Don't synchronize config_sync credentials
If the credentials are synchronized, configuring the wrong username of
password would only be detected during the next configuration update at
which point it would be too late to prevent it from affecting the whole
cluster.

Propagating functional changes to the synchronization mechanism itself
seems error prone with not that many benefits. The update interval and
timeout should be comparatively safe as they are less likely to cause
problems.