MXS-2808: Detect changes in config sync credentials
The normal way of having getters and setters for the parameters doesn't
work as well in this case as the configuration update can consist of
multiple changes and it is possible that the credentials are changed at
the same time other configuration parameters are changed. This means that
the new credentials should only be used after the current configuration
change has completed. Implementing this as a boolean reset flag seems to
be a very simpler solution for this.

As the on_set callbacks are called only when the configuration parameters
are modified, we don't need to compare the values.