MXS-2808: Track config state even with --skip-sync
The process_config() function relies on the current JSON configuration
being up to date, all successful changes that are not synchronized must
also cause it to be recalculated. This prevents them from causing the real
configuration and the JSON configuration from going out of sync.

Another solution to this would be to always recalculate the current
configuration whenever a configuration change is either made on the local
MaxScale or synced from the cluster.