MXS-3613: Add support for metadata skip
The MARIADB_CLIENT_CACHE_METADATA flag is now advertised by MaxScale and
understood by the result parsing code. Added the RCAP_TYPE_OLD_PROTOCOL
routing capability to allow the feature to be disabled for modules that do
not support the caching of metadata. All other protocol extensions will be
added under it as they are related.

Currently, the cache, masking and smartrouter modules do not support it as
they use custom code for processing the result metadata as well as the
result itself.