MXS-3578: Fix pinging after COM_CHANGE_USER
The command stored in m_reply is not updated for buffers marked as
ignorable. If the ping triggered by connection_keepalive is done right
after a COM_CHANGE_USER, the value of m_reply.command() would contain the
COM_CHANGE_USER which triggered an extra COM_CHANGE_USER to be sent. As
this code was before the code that incremented the number of ignored
replies, the reply was not ignored and triggered the debug assertion.