MXS-3621 Handle short-circuited request handling
If a service uses a filter that can short-circuit the request
handling, that is, the response can be generated immediately
and not only after a roundtrip to the server, clientReply()
will be called before routeQuery() has returned.

So that the response handling will work, the internal state
must be altered before the request handling starts, and not
after.

This is not a problem for routers, filters, etc. because things
have been arranged in the session so that also in the
short-circuited case, routeQuery() will return before
clientReply() is called. However, since the protocol resides
before the session, that arrangement is of no help here.