MXS-3623: Respond with the same auth token
Storing the token sent in the COM_CHANGE_USER will make sure the response
to the AuthSwitchRequest uses the same one that the COM_CHANGE_USER
did.

This could occur when the COM_CHANGE_USER sent to reset a persistent
connection would not complete before a COM_CHANGE_USER that used a
different set of credentials was sent by the client.