MXS-3650 Allow explicit native type
If a native type is used for holding configuration value
there will by default be an instantion of the template
class "Native" for handling the configuring at startup and
for generating a textual version of the value, if the
configuration is dumped. With this change it is possible
to use a custom type instead of the default "Native".