Fix duplication of targets in a service
This caused twice the number of connections being
opened to all servers of ReadWriteSplit.

If a service defined 'cluster=...' in the config, the
list of targets (servers) was duplicated. This was caused by
commit efb6d2d1c. The commit eludes to the possibility that
Monitor::populate_services() can be called before servers have
been added to the Monitor (at startup). Either way this commit
is valid.