Fix the pinloki_switchover test
Simplify the test to only test what is necessary.

The test failed because a slave replicating from pinloki
did not have the expected monitor state "Slave of External
 Server, Running" but just "Running". The slave was properly
replicating from pinloki, so the monitor status might be
wrong (bug) or incorrect assumptions might have beed made in
the test. In any case the test works properly and tests
what needs to be tested.