Fix service connection accounting
This is only the bug fix part, there is more. Clients connecting
in parallel could lead to more than max_connections for a service.
This was because "n_current", the db connection count was used,
not the count of client connections.

Services that are used by other services do not check the
connection count. This might be the correct behavior if enforced
at configuration time. The top level service should be the one
and only to allow max_connections.

The Target::Stats struct is a mix of client/db connection data,
they should be separated. There is also false sharing in the
usage of the struct.