Fix debug assertion on failed sescmd
If a response to a session command was expected while a query was queued,
the number of replies would be zero as the pending query is still in the
query queue. This means that the response count should only be decremented
if a response was expected and the backend was not executing a session
command.

This fixes the debug assertion encountered in the crash_on_bad_secmd test.