Fix mxs1457_ignore_deleted
The test relied on the users being updated only after the user was
created. Often the test failed as MaxScale had loaded the users right
after the user was created but before it was dropped. At that point the
authentication would still work and the test would fail as it only checks
if the connection can be created.

To fix this, the users should be refreshed before the connection is
created.