MXS-3671 Refactor query module state
Refactoring it to use worksheet states.
This keeps standalone worksheet state such as
`loading_prvw_data, prvw_data, prvw_data_request_sent_time, ...`
outside of `worksheets_arr` state. This is because
some of those states are used as a computed property in vue component.
Whenever the worksheet object in `worksheets_arr` has any properties changes,
it triggers unnecessary update as the worksheet object is used
as reactive dependencies instead of the property. By keeping it
outside the object, computed properties are cached properly.