MXS-3574: Add editing notes for documentation
The text file in the directory with the auto-generated documents hopefully
helps avoid people from editing the generated documents directly.

Added the missing step in the documentation about starting the servers.

Changed the configuration to not use syslog as it needs to be disabled for
the API documentation in version 6. Otherwise the log data example will
use syslog which isn't as readable as the maxlog version.