MXS-3669 Fix visualized graph being removed
When adding a new worksheet, the graph of previous tab is removed since
the `resultSets` watcher value is computed by using standalone worksheet state,
so when the states change, it clears data and removes the graph.

This uses the imperative `$watch` to dynamically watch/unwatch `resultSets`
when active worksheet changes.

This also removes obsolete `TODO`