Increase CTest timeout to 1200 seconds
This should help avoid the case where the first test will fail due to a
CTest timeout because the VM creation takes too long. One example of this
is mxs1506_delayed_retry which is a relatively long test that is the first
one to be executed.

Ideally, the test startup time wouldn't be included in the timeout but
CTest has no way of knowing what part of the execution is meant to be a
test and what isn't. Since the test already tracks timeouts by itself, the
CTest timeout isn't that critical.