MXS-3594: Make regexfilter runtime configurable
Using the mxs::config types in the filter would probably result in code
that's very similar to the current code with the exception that it would
introduce a global lock.

Due to the possibility of the log file changing at runtime, it had to be
changed from a global FILE handle into individual handles. This prevents
the need to synchronize the access to the file stream but it does increase
the number of files that are opened. Both of the implementations still
rely on the operating system to guarantee the atomicity of the writes.

This also fixes the bug where the post_configure method didn't override
the method it was supposed to override.