Document nosqlprotocol creation with MaxCtrl
Due to the way the nested parameters work, the parameters must be passed
as a single object. As this is currently the only way to pass them to the
REST API, it should be documented.

Some smartness in MaxCtrl might make it possible to pass these in a
similar manner to how they are passed in the configuration.If this is
implemented, the documentation must be updated to use the more
user-friendly version.