MXS-3687: Add prefixes for Specifications
The prefix allows parameters to be explicitly declared in their own
"namespace". The use of a prefix in a Specification causes the
configuration processing code to assume that all parameters declared in
the Specification are to be prefixed with the given prefix.

This fixes the serialization of the nosqlprotocol which is currently the
only module that uses parameters that have a prefix.