MXS-3687: Invert nested parameter check
The parameter check is now done in terms of the parameters_with_params
which allows missing parameters to be detected. As a result of this, all
processed parameters have to be removed from the nested parameter
container to detect nested parameters for unknown modules.

As the listener parameters are validated in a single step, the container
for the unrecognized parameters is not needed.