MXS-3704: Don't send SESSION_TRACK for XPand clusters
The capability should not be advertised if it isn't supported by the
backend nodes. This is mainly a problem with XPand that doesn't support
either the extended MariaDB capabilities or the later additions into the
basic capabilities.

Other capabilities advertised by MaxScale might also need to be adjusted
whenever XPand nodes are in use but the main problem was with DBeaver and
SESSION_TRACK.