MXS-3299: Add alias support to binlogrouter
Adding support for field aliases is not immensely complex and it improves
the degree to which the binlogrouter can process SQL. The next step would
be to introduce some constant variables that make the binlogrouter seem
more like an actual database.

The added grammar does allow for a non-standard trailing AS keyword but as
it doesn't conflict with the standard SQL and it keeps the grammar
extremely simple, it is a reasonable compromise for a solution that is
added to a bugfix release.