Preload users for crash_out_of_files
This makes sure the test won't fail due to the users not being
available. This would sometimes happen when one of the services manages to
create enough connections to block out other services. This would also
block the user loading from creating a connection which in turn would
eventually cause the host to get blocked as the authentication would keep
failing.