MXS-3681 Fix `visualize-sidebar` is not shown properly
Toggling the visualized config button shows the sidebar as expected
but after refreshing the page, the sidebar is not shown. This
is because the function handles showing the sidebar isn't called when
the component is "activated".

This calls `handleSetVisSidebar` in the next tick after result pane dimension
has been updated. This also make the chart-container pane show only when
the chart has been fully configured.