Add delayed_call debug assertions
The delayed_call function of a worker must only be called on the thread
that actually runs the worker. The new debug assertions in
Worker::delayed_call and Worker::cancel_delayed_call make sure the code
behaves as expected.

Adding these assertions revealed that the MainWorker violated this
assumption by adding delayed calls in its constructor. Although this was a
perfectly safe thing to do, it is better to do this work in
MainWorker::pre_run where the assertions will hold.