MXS-3681 Fix memory leak issue in query editor (part1)
All worksheet states including schemas tree data and result set data
are stored in localStorage and in memory to keep query editor's state
when refreshing the browser and switching between worksheets. However,
when it reaches the size limit of localStorage, the browser is frozen.
Hence, these data shouldn't be kept in localStorage.

This moves schemas tree data's and related states outside of `worksheets_arr`
and keep it in memory only. To keep the state when browser is refreshed,
this refactors all fetching actions which are used to populate the schemas tree by
utilizing `expanded_nodes` values that is stored in localStorage.