MXS-3037: Add 'list queries' to MaxCtrl
The command allows all queries that are still active to be listed. The
format is roughly equivalent to the SHOW PROCESSLIST output of the MariaDB
server except that it doesn't show idle connections as that's what the
'list sessions' command is for.

The actual server where the query is being executed on would be a good
addition but this information is currently not available in MaxScale.