MXS-3713 Prepare for packet command classes
Currently it is arranged so that OP_QUERY can be handled
by command classes geared for OP_MSG, but now that other
packet types must be handled it would become to much of
a mess to continue on that path. Instead common
functionality will be factored out and there will be a
separate command class (hierarchy in the case of OP_MSG)
for each packet type.