Fix JSON equality comparison
If two undefined JSON objects are compared in JavaScript, they compare
equal.

The behavior should be the same in MaxScale as there are some benefits:
checking if two JSON objects contain the same value for both keys can be
done by comparing the values regardless of whether they exist. As long as
undefined values compare equal, the end result is the same. This makes
detecting changes in JSON objects much easier.