Make waiting smarter in mxs1687_avro_ha
Waiting for a predetermined amount of seconds without checking that the
other monitor actually acquired the locks hasn't proven to be very
reliable. A more reliable approach is to wait until the desired monitor
actually gains the locks. This seems to have improved the local stability
of the test.

There was also a problem where the avrorouter would be slow to react to
changes if there were too many events in the binlog right at the start. To
prevent this problem and to provide a stable starting point, the
replication can be reset before the test is started.

Also added the comments as output to the test, this should explain which
part the test failed without having to look into the source code.