MXS-3681 Fix `active_conn_state`
Simply checking for valid connection ids is not accurate.
If a worksheet was connected to a resource and the token is expired
or deleted, but app's state still persists the value, this redirects
user to 404.

This fixes the value of `active_conn_state` and show snackbar to indicate
user that the connection is not found.