MXS-3710: Make namedserverfilter runtime configurable
Storing a shared pointer in the WorkerGlobal allows the same compiled
regular expressions to be used by all sessions. This retains the low
memory overhead of the current implementation.

The filter currently doesn't do pre-validation of the parameters which
causes the configuration to be changed before it is known to be
good. Right now this only seems to cause problems in the diagnostic output
where the bad configuration is shown. This needs to be handled by a
post_validate step done by the filter's Specification.