MXS-3710: Validate namedserverfilter parameters
The post_validate step now prevents the actual configuration from being
modified if the generated configuration would not be valid.

Unfortunately, by moving the validation away from the actual
configuration, the patterns are compiled twice during a valid
configuration attempt. As this only happens on startup and during
alterations, it's not that relevant as it's the MainWorker thread that
does the extra work. Not the prettiest solution but it's what we have.