MXS-3583 Add [Create|Drop|List]Indexes commands
Currently it is not actually possible to create indexes.
However, as indexes relates to the performance and not to
the actual functionality, claiming that index creation is
possible is likely to be better than just erroring.

TODO: The mechanism for creating the table/collection, if
not present, should be moved to OpMsgCommand as many commands
depend on that.