MXS-3720: Disable idle pooling on COM_STMT_PREPARE
The COM_STMT_EXECUTE command contains metadata that describes the data
sent in the command. The metadata can be omitted if the same statement is
executed with the same types of parameters. This is an optimization that
reduces the amount of data sent over the network.

Currently, only readwritesplit handles this correctly. It does this by
storing a copy of the metadata and adding it once for each
COM_STMT_EXECUTE executed on a backend.

As the metadata must be added in the first COM_STMT_EXECUTE, the failure
to add it will cause the command to fail. To make things worse, there is a
bug (MDEV-19811) in the server's prepared statement code that causes it to
crash if the metadata is not sent in the first COM_STMT_EXECUTE.

The way that idle_session_pool_time relates to this is that even if a
COM_STMT_EXECUTE was once correctly executed with the metadata added to
it, the act of putting the connection into the pool causes that metadata
to be lost. As this metadata is currently stored only in readwritesplit,
the pooling cannot be safely done.

The best way to solve this would be to move the metadata cache from
readwritesplit into the client protocol. This would allow the backend
protocol to append the metadata when needed and the connections could be
safely put into the pool and taken out of it.

For the time being, the most effective solution to this problem is to
simply disable the pooling if a COM_STMT_PREPARE is seen.