Move PS metadata cache to MariaDB protocol
The metadata cache is more efficient and slightly simpler to implement at
the protocol level than it is at the router level. This also has the added
benefit of automating the metadata addition if a connection is taken from
the pool.

The code that was added is largely the same that was in readwritesplit
except that it was split into two parts. The metadata extraction can
always be done by the client protocol as it's the first one to receive
it. The addition of the metadata to the COM_STMT_EXECUTE is done by the
backend protocol as it is the only one who know what was actually sent to
the server.

This commit also fixes the problems seen with COM_STMT_EXECUTE and
idle_session_pool_time but due to the still unknown extent of the problems
it introduces, it is better left disabled.