MXS-3731 Prevent used connection from being used
Once a connection is used by a worksheet, it will be disabled
to be selected in other worksheets. Although, it still can be
disconnected. e.g. `server-0` is connected in `worksheet-1`,
in `worksheet-2`, user can disconnect `server-0` of `worksheet-1`.
This makes things easier for user to disconnect connections.

This also removes the editable feature of worksheet tab name. Worksheet
tab will now use connected resource name as tab name, if user has not
selected connection to be used within that worksheet, "WORKSHEET" placeholder
name will be used.