MXS-3475 Add option to add custom user to user account manager
This is required when using group mapping with XPand. Xpand does not
support either PAM or anonymous users, so detection of the anonymous
user will normally never trigger. Such a user needs to be added
manually.

This commit adds several authenticator options to define such a user.
Using the 'authenticator_options'-setting for this is unideal, but
a quick solution is required.