MXS-3657: Fix regex options in ccrfilter
Since the ccrfilter was the only module that used RegexValue::set_options,
this change does not affect other modules.

This is different in version 6 where most modules that have regular
expressions have upgraded to the built-in Regex type, either the
mxb::Regex variant or the derived version of it in the mxs::config
namespace.