MXS-3732 Refactor `db_tree` state
`db_tree` state was using to store db list schema of a worksheet. When
switching to a new worksheet, there would be requests to populate the
schema tree. However, when working with parallel requests of different connections,
this causes prallel requests of the same connection .e.g. user is querying
result data in worksheet 1 which would take 20 seconds, then user switches
to worksheet 2 and execute another queries then immediately switch back to
previous worksheet. There would be requests to populate the schema tree but encounters
parallel requests of the same connection id.

This refactors `db_tree` state to `db_tree_map` to store worksheet's schema tree
data in memory. Requests to populate schema tree will be sent once when a worksheet
is attached to a connection.