MXS-3738: Rewrite `show dbusers`
The user account manager data is now contained in the service
resource. This is more in line with how the data is actually stored: all
listeners share the same user account data which is stored in the top
level service.

Most of the `show dbusers` command was rewritten to better format this
data. Since the command has been broken for all of 2.5, this shouldn't
cause any problems with backwards compatibility.