MXS-3732 Move `is_validating_conn` from localStorage
Since `validatingConn` action is called one time when query page is created
and it's not binded to a worksheet, it can be stored in memory. Apart from that,
`active_wke_id` is an empty string when the action is called, so binding `active_wke_id`
to `SET_IS_VALIDATING_CONN` mutation is useless.

This also clears `curr_cnct_resource` if current connected resource id is not valid
while validating connections.