Define static member variables for ODR-use
From: https://en.cppreference.com/w/cpp/language/static

> If a const static data member is odr-used, a definition at namespace
  scope is still required, but it cannot have an initializer.

The two constant variables are ODR-used due to them being passed as values
to std::min and std::max which take their values by reference.

C++17 removes this limitation with the addition of inline variables. This
would make these constants work the way they are intended to work but as
we compile MaxScale in C++14, we still have to declare them.

This commit also fixes the formatting errors in worker.cc.