Fix `call command` output
If the response was the green OK string that is returned for a HTTP 200
OK, it would be converted into a JSON string which caused the output
formatting to be lost. For this, checking the type of the response is
enough as no other response will return a plain string.

Additionally, the output from commands that did produce it contained
irrelevant data. As the module-generated data is always stored in the meta
field, the output from that can be returned instead of the whole object.