MXS-3308: Prevent global options from being redefined
This will prevent the odd errors that occur when the arguments get
converted into arrays if they are used multiple times. This won't solve
the case where multiple arguments are passed in non-interactive mode but
this is a smaller problem as that's easily visible and not something
that's implicitly added by MaxCtrl.