MXS-3728: Catch exceptions from delayed calls
The same function is called from two different contexts, one where the
thrown exceptions are caught and another where they weren't.

This does reveal a problem: the errors won't be returned to the client if
they are generated after the original Reader object has been
constructed. A separate "abort" handler would be needed for this or the
delayed call would have to be moved one level up into the Pinloki session.