MXS-3732 Add `catchAction` and `actionName` as parameters of `queryingActionWrapper`
When using `queryingActionWrapper` as a wrapper function,
if the action calls inside it also has `try...catch` statement and
encounters error, the inner catch block is ignored and the outer catch
is called. So, some actions needs to reset loading state when encounter
error won't run.

This adds `catchAction` as a callback function and use `actionName`
to create logger.