MXS-3709: Make masking filter configurable at runtime
Storing the compiled rules in the configuration conveniently packages it
into something the filter session can copy during construction. This
freezes the whole configuration of a filter which prevents modifications
done after the session has started from affecting the filter session.

This also should be faster than using a global lock but that does come at
the cost of extra memory usage.