MXS-3740 Make shortcut keys of editor available in page context
The shortcut keys were made to be available only when user focuses the cursor
on the editor area.

This uses `vue-shortkey` plugin to add those shortcut keys to be also
available in page context. This still keeps events registered and emitted
from the `query-editor` as it should be shown in the Command Palette.
In addition, when the editor is on focus, all other key events outside
the editor are blocked which including key events of vue-shortkey.
So, key events on page context and editor context are needed.