MXS-3681 Fix sidebar schemas isn't reloaded when creating new connection
If current worksheet has been bound to a connection already and
user creates new connection, the `initialFetch` action wasn't called.
This is because there is a watcher that watches on `curr_cnct_resource` value,
even when the value changes, it won't dispatch the action so as to prevent uncessary
fetches when switching between worksheets, it checks for the existence of
schemas tree data.

This checks for existing connection of active worksheet before creating new
connection. If current worksheet was bound to a connection and successful
connecting to new connection, it dispatches `initalFetch` to populate new
schemas tree data.