MXS-3718: Fix WatchdogNotifier semaphore waits
The POSIX semaphore uses CLOCK_REALTIME for sem_timedwait which makes it
vulnerable to changes in the system time. The standard
std::condition_variable would be the correct replacement for this but due
to a bug in GCC 9 and older , it cannot be used as it also uses the system
clock for wait_for. As a workaround, a partial implementation of the
std::condition_variable has to be done in maxbase.

A few minor modifications to the Ticker class allow the double semaphore
approach to be replaced with a single condition variable. Since the
incrementing of the client count was already protected by a lock, it is
natural to use it also in the actual thread that does the notifying.