MXS-3681 Fix async issue when calling `initialFetch`
When user closes the browser and if auth token or connection token
are expired after that, then there is no `logout` action being
dispatched. As a result, `resetState` function won't be called,
so persisted states in localStorage will not be cleared. When
user enters query page, `validatingConn` action is dispatched
but the `initalFetch` action is called before that already which
leads user to 404 page.

The `initialFetch`  action should be dispatched only after finish validating
connections.