MXS-3331: Use SO_REUSEADDR with bound client sockets
With SO_REUSEPORT configured for client sockets, local sockets in
TIME_WAIT state can be reused (similar to how listening sockets behave)
which allows more outbound connectios to be used. This only matters when
using local_address as the call to bind() is what fails, not the connect()
call.

Examples of similar problems can be found in these two articles:

  https://blog.heroku.com/sockets-in-a-bind
  https://idea.popcount.org/2014-04-03-bind-before-connect/

The articles also contains some hints for a few optimizations that could
be done for local_address, namely the use of multiple local_address values
and the use of the IP_BIND_ADDRESS_NO_PORT socket option.