MXS-3742 Add `max-content` for width attr of row count column
Row count column .i.e. `#` column can have `max-content` as
a value for width. So the `#` column won't stretch its width to be
equal with others columns.

This refactors `resetHeaderWidth` to assign the value of `width`
attribute from `header` object to `headerWidthMap` directly without
using `$refs` to the header element as it's not realiable. As a result,
the value of each key in `headerWidthMap` can be either string or number.