MXS-3583 Use JSON_QUERY with JSON_CONTAINS
And not JSON_EXTRACT. JSON_QUERY returns an array or a document,
but not a value (NULL in that case). JSON_CONTAINS operates on
arrays and documents, and does the right thing with nulls but
cant deal with values, in which case the result may be a strange
syntax error.