MXS-3742 Add `groupBy` feature to `virtual-scroll-table`
`groupBy` is useful when user wants to group rows have the
same value by a column.

`groupBy` feature  is enabled by providing `groupable` key to
a header object.

This adds `group` text as button to click which is placed next to
sort icon. When clicking it, it adds another row with a button
to toggle the expansion/collapsion of grouped rows and a button
to ungroup. The header grouped row is presented with column name to be grouped
and corresponding value of that column. e.g. If users group by `connection_name`,
the header grouped row is presented as follows: `connection_name: server_1`.