MXS-3742 Add checkbox to delete history/favorite queries
User can select one or all rows to delete

This adds checkbox to `virtual-scroll-table` and adds
delete button to `result-data-table`.

With current implementation, Virtual-scroll-table uses 2d array to
render table rows so when toggling the visibility of columns,
it's not possible to detect which rows to be deleted or selected.
So there is no other way but to deep compare original history/favorite
rows with selected rows. As a result, columns toggle needs to be
disabled and reset when selecting rows.